+++ Tests/Ifrit_Levenstein/Levenstein_TestsCore.swift
@@ -1,8 +1,9 @@
+
 import Foundation
 import XCTest
 @testable import Ifrit
 
-final class Levenstain_TestsCore: XCTestCase {
+final class Levenstein_TestsCore: XCTestCase {
     func testMatch() {
         let actual = bestMatch(query: "ff", input: "FuzzyFind")
         let expected = Alignment(
@@ -18,25 +19,25 @@ final class Levenstain_TestsCore: XCTestCase {
         )
         XCTAssertEqual(actual, expected)
     }
-
+    
     func testContiguousCharactersHaveHigherScore() {
         let a1 = bestMatch(query: "pp", input: "pickled pepper")!
         let a2 = bestMatch(query: "pp", input: "Pied Piper")!
         XCTAssertTrue(a1.score > a2.score)
     }
-
+    
     func testCharactersAtBeginningOfWordsHaveHigherScore() {
         let a1 = bestMatch(query: "pp", input: "Pied Piper")!
         let a2 = bestMatch(query: "pp", input: "porcupine")!
         XCTAssertTrue(a1.score > a2.score)
     }
-
+    
     func testCamelCaseHumpsHaveHigherScore() {
         let a1 = bestMatch(query: "bm", input: "BatMan")!
         let a2 = bestMatch(query: "bm", input: "Batman")!
         XCTAssertTrue(a1.score > a2.score)
     }
-
+    
     func testFirstLettersOfWordsHaveHigherScore() {
         let a1 = bestMatch(query: "bm", input: "Bat man")!
         let a2 = bestMatch(query: "bm", input: "Batman")!